requirements.txt
PySide6>=6.6.0
wdtagger>=0.1.0
timm>=0.9.0
onnxruntime>=1.15.0
scipy>=1.10.0
Pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0
pyinstaller>=6.0.0

